Reorganized settings and styled the settings button
