+++ packages/inspect-ui/app/app.tsx
@@ -1,40 +1,22 @@
-import React, { useState } from "react";
-import { RouteObject, useRoutes } from "react-router-dom";
 import { Slice } from "@effection/atom";
 import { InspectState } from "@effection/inspect-utils";
-import {
-  SettingsMenu,
-  SettingsContext,
-  DEFAULT_SETTINGS,
-} from "./components/settings";
+import AppBar from "@material-ui/core/AppBar";
+import Container from "@material-ui/core/Container/Container";
+import CssBaseline from "@material-ui/core/CssBaseline";
+import React, { useState } from "react";
+import { RouteObject, useRoutes } from "react-router-dom";
+import { SettingsButton } from "./components/settings-button";
+import { SettingsForm } from "./components/settings-form";
+import { DEFAULT_SETTINGS, SettingsContext } from "./context";
+import { useStyles } from "./hooks/use-styles";
 import { TaskPage } from "./task-page";
 import { TaskTreePage } from "./task-tree-page";
-import AppBar from "@material-ui/core/AppBar";
-import { makeStyles } from "@material-ui/core/styles";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import Container from "@material-ui/core/Container/Container";
 
 export type InspectStateSlice = Slice<InspectState>;
 
 type AppProps = {
   slice: InspectStateSlice;
 };
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: "flex",
-  },
-  content: {
-    flexGrow: 1,
-    height: "100vh",
-    overflow: "auto",
-  },
-  appBarSpacer: theme.mixins.toolbar,
-  container: {
-    paddingTop: theme.spacing(4),
-    paddingBottom: theme.spacing(4),
-  },
-}));
 
 export function App({ slice }: AppProps): JSX.Element {
   let classes = useStyles();
@@ -61,8 +43,10 @@ export function App({ slice }: AppProps): JSX.Element {
     <SettingsContext.Provider value={{ settings, setSettings }}>
       <div className={classes.root}>
         <CssBaseline />
-        <AppBar position="absolute">
-          <SettingsMenu />
+        <AppBar position="absolute" className={classes.appBar}>
+          <SettingsButton>
+            <SettingsForm />
+          </SettingsButton>
         </AppBar>
         <main className={classes.content}>
           <div className={classes.appBarSpacer}>
