+++ packages/inspect-ui/app/hooks/use-styles.tsx
@@ -1,0 +1,22 @@
+import { makeStyles } from "@material-ui/core/styles";
+
+export const useStyles = makeStyles((theme) => ({
+  root: {
+    display: "flex",
+  },
+  content: {
+    flexGrow: 1,
+    height: "100vh",
+    overflow: "auto",
+    paddingTop: theme.spacing(3),
+  },
+  appBar: {
+    display: "flex",
+    alignItems: "end",
+  },
+  appBarSpacer: theme.mixins.toolbar,
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4),
+  },
+}));
