+++ packages/inspect-ui/app/utils/search-string.tsx
@@ -1,0 +1,22 @@
+import { stringify } from "query-string";
+export type SearchParams = {
+  [key: string]: string | boolean | number | undefined;
+};
+
+export function searchString(parts: SearchParams): string {
+  // remove keys with undefined to allow
+  let filtered = Object.keys(parts).reduce<SearchParams>((acc, key) => {
+    if (parts[key] === undefined) {
+      return acc;
+    } else {
+      acc[key] = parts[key];
+      return acc;
+    }
+  }, {});
+
+  if (Object.keys(filtered).length === 0) {
+    return "";
+  } else {
+    return `?${stringify(filtered)}`;
+  }
+}
